backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
langgraph==0.2.28
fastembed==0.3.6
faiss-cpu==1.8.0
supabase==2.7.4
pymupdf==1.24.10
tavily-python==0.5.0
python-multipart==0.0.12
pydantic==2.9.2
numpy==1.26.4
httpx==0.27.2
